% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grayleafspot-package.R
\docType{package}
\name{grayleafspotr-package}
\alias{grayleafspotr}
\alias{grayleafspotr-package}
\title{Gray Leaf Spot Package}
\description{
Self-contained R tools for gray leaf spot analysis, raw exports, and plotting.
The package is intended to be used from RStudio and developed with devtools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rotsl.github.io/grayleafspotr}
  \item \url{https://github.com/rotsl/grayleafspotr}
  \item Report bugs at \url{https://github.com/rotsl/grayleafspotr/issues}
}

}
\author{
\strong{Maintainer}: Rohan R \email{phonics-tiffs1i@icloud.com} (\href{https://orcid.org/0009-0005-9225-1775}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Rohan R \email{phonics-tiffs1i@icloud.com} (\href{https://orcid.org/0009-0005-9225-1775}{ORCID}) [copyright holder]
}

}
\keyword{internal}
