% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_feature_heatmap}
\alias{plot_feature_heatmap}
\title{Pearson correlation heatmap of numeric morphology features}
\usage{
plot_feature_heatmap(x)
}
\arguments{
\item{x}{A \code{grayleafspot_run} object or data accepted by
\code{\link[=as_grayleafspot_growth_data]{as_grayleafspot_growth_data()}}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Pearson correlation heatmap of numeric morphology features
}
