% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_growth_roughness}
\alias{plot_growth_roughness}
\title{Plot relative growth rate and edge roughness}
\usage{
plot_growth_roughness(x)
}
\arguments{
\item{x}{A \code{grayleafspot_run} object or data accepted by
\code{\link[=as_grayleafspot_growth_data]{as_grayleafspot_growth_data()}}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot relative growth rate and edge roughness
}
