% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_stress_remodeling}
\alias{plot_stress_remodeling}
\title{Plot crack coverage and crack count over time}
\usage{
plot_stress_remodeling(x)
}
\arguments{
\item{x}{A \code{grayleafspot_run} object or data accepted by
\code{\link[=as_grayleafspot_growth_data]{as_grayleafspot_growth_data()}}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot crack coverage and crack count over time
}
