% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_texture_organization}
\alias{plot_texture_organization}
\title{Plot texture entropy and center-to-edge intensity delta over time}
\usage{
plot_texture_organization(x)
}
\arguments{
\item{x}{A \code{grayleafspot_run} object or data accepted by
\code{\link[=as_grayleafspot_growth_data]{as_grayleafspot_growth_data()}}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot texture entropy and center-to-edge intensity delta over time
}
